Use paths and reorganize all components
