--- fe/callisto/old-package.json
@@ -1,70 +1,0 @@
-{
-  "name": "callisto",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": { 
-    "@emotion/react": "^11.13.5",
-    "@emotion/styled": "^11.13.5",
-    "@mui/icons-material": "^6.1.9",
-    "@mui/material": "^6.1.9",
-    "@react-oauth/google": "^0.12.1",
-    "@react-three/drei": "^9.119.0",
-    "@react-three/fiber": "^8.17.10",
-    "@react-three/postprocessing": "^2.16.3",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/color": "^3.0.6",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.96",
-    "@types/react": "^18.2.79",
-    "@types/react-dom": "^18.2.25",
-    "axios": "^1.7.7",
-    "color": "^4.2.3",
-    "crypto-js": "^4.2.0",
-    "http-status-codes": "^2.3.0",
-    "rc-collapse": "^3.9.0",
-    "react": "^18.2.0",
-    "react-accessible-accordion": "^5.0.0",
-    "react-dom": "^18.2.0",
-    "react-icons": "^5.3.0",
-    "react-scripts": "^3.0.1",
-    "react-tooltip": "^5.28.0",
-    "react-use-googlelogin": "^0.12.5",
-    "rsuite": "^5.73.0",
-    "three": "^0.170.0",
-    "typescript": "^4.9.5",
-    "web-vitals": "^2.1.4",
-    "zustand": "^4.5.2"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@types/crypto-js": "^4.2.2",
-    "@types/three": "^0.163.0",
-    "buffer": "^6.0.3",
-    "process": "^0.11.10"
-  }
-}
