+++ fe/callisto/src/components/controls/ShipComputer.tsx
@@ -9,11 +9,11 @@ import {
   POSITION_SCALE,
   ViewContext,
   ViewMode,
-} from "./Universal";
+} from "lib/universal";
 
-import {setPlan, setCrewActions} from "./ServerManager";
-import {ActionContext, SensorState, SensorAction, newSensorState} from "./Actions";
-import {EntitySelectorType, EntitySelector} from "./lib/EntitySelector";
+import {setPlan, setCrewActions} from "lib/serverManager";
+import {ActionContext, SensorState, SensorAction, newSensorState} from "components/controls/Actions";
+import {EntitySelectorType, EntitySelector} from "lib/EntitySelector";
 
 // Distance in km for standoff from another ship.
 const DEFAULT_SHIP_STANDOFF_DISTANCE: number = 10;
