+++ fe/callisto/src/components/controls/WeaponUse.tsx
@@ -1,5 +1,5 @@
 import React, {useContext, useState, useMemo, useEffect} from "react";
-import {findRangeBand} from "./lib/Util";
+import {findRangeBand} from "lib/Util";
 import {
   Entity,
   ShipDesignTemplate,
@@ -8,24 +8,24 @@ import {
   DesignTemplatesContext,
   Ship,
   SHIP_SYSTEMS,
-} from "./Universal";
-import {EntitySelector, EntitySelectorType} from "./lib/EntitySelector";
-import {FireState, ActionContext, PointDefenseState} from "./Actions";
+} from "lib/universal";
+import {EntitySelector, EntitySelectorType} from "lib/EntitySelector";
+import {FireState, ActionContext, PointDefenseState} from "components/controls/Actions";
 
 // Icons for each type of weapon
-import {ReactComponent as Turret1} from "./assets/icons/turret1.svg";
-import {ReactComponent as Turret2} from "./assets/icons/turret2.svg";
-import {ReactComponent as Turret3} from "./assets/icons/turret3.svg";
-import {ReactComponent as Barbette} from "./assets/icons/barbette.svg";
-import {ReactComponent as SmallBay} from "./assets/icons/bay-s.svg";
-import {ReactComponent as MediumBay} from "./assets/icons/bay-m.svg";
-import {ReactComponent as LargeBay} from "./assets/icons/bay-l.svg";
+import {ReactComponent as Turret1} from "assets/icons/turret1.svg";
+import {ReactComponent as Turret2} from "assets/icons/turret2.svg";
+import {ReactComponent as Turret3} from "assets/icons/turret3.svg";
+import {ReactComponent as Barbette} from "assets/icons/barbette.svg";
+import {ReactComponent as SmallBay} from "assets/icons/bay-s.svg";
+import {ReactComponent as MediumBay} from "assets/icons/bay-m.svg";
+import {ReactComponent as LargeBay} from "assets/icons/bay-l.svg";
 
 // Icons to show fire states.
-import {ReactComponent as RayIcon} from "./assets/icons/laser.svg";
-import {ReactComponent as MissileIcon} from "./assets/icons/missile.svg";
+import {ReactComponent as RayIcon} from "assets/icons/laser.svg";
+import {ReactComponent as MissileIcon} from "assets/icons/missile.svg";
 import {Tooltip} from "react-tooltip";
-import {vectorDistance} from "./lib/Util";
+import {vectorDistance} from "lib/Util";
 
 // Consistent set of colors for both type of weapons and fire states.
 const WEAPON_COLORS: {[key: string]: string} = {
