+++ fe/callisto/src/components/space/Effects.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 import { useContext } from "react";
 import { animated, useSpring } from "@react-spring/three";
-import { scaleVector } from "./lib/Util";
-import { SCALE, EntitiesServerContext } from "./Universal";
-import { GrowLine } from "./lib/Util";
+import { scaleVector } from "lib/Util";
+import { SCALE, EntitiesServerContext } from "lib/universal";
+import { GrowLine } from "lib/Util";
 
 const SHIP_IMPACT = "ShipImpact";
 const EXHAUSTED_MISSILE = "ExhaustedMissile";
