+++ fe/callisto/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "baseUrl": "./",
+    "baseUrl": "./src",
     "target": "es2015",
     "lib": [
       "dom",
@@ -8,6 +8,9 @@
       "esnext"
     ],
     "paths": {
+      "components/*": ["components/*"],
+      "lib/*": ["lib/*"],
+      "assets/*": ["assets/*"],
       "three/addons/controls/*": ["node_modules/@types/three/examples/jsm/controls/*"],
       "three/addons/loaders/*": ["node_modules/@types/three/examples/jsm/loaders/*"],
     },
